[refactor] 도메인 이름 변경
문제 사항 :
 - 도메인의 이름이 bootcamp보다는 certification이 더 적절하다고 판단

변경 사항 :
 - 도메인 이름 변경
 - bootcamp -> certification
