--- src/main/java/com/fasttime/domain/bootcamp/dto/request/CertificationRequestDTO.java
@@ -1,9 +1,0 @@
-package com.fasttime.domain.bootcamp.dto.request;
-
-public record CertificationRequestDTO(
-    String bootcampName,
-    String image,
-    String content
-) {
-
-}
