--- src/main/java/com/fasttime/domain/bootcamp/dto/request/RejectionRequestDTO.java
@@ -1,9 +1,0 @@
-package com.fasttime.domain.bootcamp.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RejectionRequestDTO(
-    @NotBlank(message = "거절 사유는 반드시 입력해야 합니다.") String rejectionReason
-) {
-
-}
