--- src/main/java/com/fasttime/domain/bootcamp/entity/BootCamp.java
@@ -1,46 +1,0 @@
-package com.fasttime.domain.bootcamp.entity;
-
-import com.fasttime.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "bootcamp")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class BootCamp extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-
-    @Column(columnDefinition = "TEXT")
-    private String image;
-
-    @Column(name = "government_funded")
-    private boolean governmentFunded;
-
-    private String organizer;
-
-    private String website;
-
-    @Column(columnDefinition = "TEXT")
-    private String course;
-
-}
