--- src/main/java/com/fasttime/domain/bootcamp/entity/Certification.java
@@ -1,85 +1,0 @@
-package com.fasttime.domain.bootcamp.entity;
-
-import com.fasttime.domain.member.entity.Member;
-import com.fasttime.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Getter
-public class Certification extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "bootcamp_name", nullable = false)
-    private String bootcampName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @Column(name = "image", columnDefinition = "TEXT")
-    private String image;
-
-    @Column(name = "content", columnDefinition = "TEXT")
-    private String content;
-
-    @Column(name = "withdrawal_reason", columnDefinition = "TEXT")
-    private String withdrawalReason;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "bootcamp_id")
-    private BootCamp bootCamp;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private CertificationStatus status = CertificationStatus.PENDING;
-
-    @Column(name = "rejection_reason", columnDefinition = "TEXT")
-    private String rejectionReason;
-
-    public void setWithdrawalReason(String withdrawalReason) {
-        this.withdrawalReason = withdrawalReason;
-    }
-
-    public void setRejectionReason(String rejectionReason) {
-        this.rejectionReason = rejectionReason;
-    }
-
-    public void setStatus(CertificationStatus status) {
-        this.status = status;
-    }
-
-    public void setBootCamp(BootCamp bootCamp) {
-        this.bootCamp = bootCamp;
-    }
-
-    public void softDelete() {
-        delete(LocalDateTime.now());
-    }
-
-    @Override
-    public void restore() {
-        super.restore();
-    }
-}
