--- src/main/java/com/fasttime/domain/bootcamp/entity/CertificationStatus.java
@@ -1,5 +1,0 @@
-package com.fasttime.domain.bootcamp.entity;
-
-public enum CertificationStatus {
-    PENDING, APPROVED, REJECTED, WITHDRAW
-}
