--- src/main/java/com/fasttime/domain/bootcamp/exception/CertificationNotFoundException.java
@@ -1,13 +1,0 @@
-package com.fasttime.domain.bootcamp.exception;
-
-import com.fasttime.global.exception.ApplicationException;
-import com.fasttime.global.exception.ErrorCode;
-
-public class CertificationNotFoundException extends ApplicationException {
-
-    private static final ErrorCode ERROR_CODE = ErrorCode.CERTIFICATION_NOT_FOUND;
-
-    public CertificationNotFoundException() {
-        super(ERROR_CODE);
-    }
-}
