--- src/main/java/com/fasttime/domain/bootcamp/repository/BootCampRepository.java
@@ -1,8 +1,0 @@
-package com.fasttime.domain.bootcamp.repository;
-
-import com.fasttime.domain.bootcamp.entity.BootCamp;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BootCampRepository extends JpaRepository<BootCamp, Long> {
-    boolean existsByName(String name);
-}
