+++ src/main/java/com/fasttime/domain/certification/dto/request/CertificationRequestDTO.java
@@ -1,0 +1,9 @@
+package com.fasttime.domain.certification.dto.request;
+
+public record CertificationRequestDTO(
+    String bootcampName,
+    String image,
+    String content
+) {
+
+}
