+++ src/main/java/com/fasttime/domain/certification/entity/CertificationStatus.java
@@ -1,0 +1,5 @@
+package com.fasttime.domain.certification.entity;
+
+public enum CertificationStatus {
+    PENDING, APPROVED, REJECTED, WITHDRAW
+}
