+++ src/main/java/com/fasttime/domain/certification/exception/CertificationBadRequestException.java
@@ -1,0 +1,17 @@
+package com.fasttime.domain.certification.exception;
+
+import com.fasttime.global.exception.ApplicationException;
+import com.fasttime.global.exception.ErrorCode;
+
+public class CertificationBadRequestException extends ApplicationException {
+
+    private static final ErrorCode ERROR_CODE = ErrorCode.CERTIFICATION_BAD_REQUEST;
+
+    public CertificationBadRequestException() {
+        super(ERROR_CODE);
+    }
+
+    public CertificationBadRequestException(String message) {
+        super(ERROR_CODE, message);
+    }
+}
