+++ src/main/java/com/fasttime/domain/member/entity/Member.java
@@ -1,6 +1,6 @@
 package com.fasttime.domain.member.entity;
 
-import com.fasttime.domain.bootcamp.entity.BootCamp;
+import com.fasttime.domain.certification.entity.BootCamp;
 import com.fasttime.global.common.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
